backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic==2.11.9
pydantic-settings==2.4.0
SQLAlchemy==2.0.32
python-dotenv==1.0.1
alembic==1.13.2
httpx==0.28.1
python-multipart==0.0.9
google-cloud-documentai==3.6.0
google-cloud-aiplatform==1.71.1
supabase==2.18.1
google-generativeai==0.8.5